refactor: status domain/status로 이동
