+++ src/main/java/com/gamemoonchul/domain/status/Oauth2Status.java
@@ -1,4 +1,4 @@
-package com.gamemoonchul.config.oauth;
+package com.gamemoonchul.domain.status;
 
 import com.gamemoonchul.common.status.ApiStatusIfs;
 import lombok.AllArgsConstructor;
